+++ web/src/components/layout/nav.tsx
@@ -182,6 +182,11 @@ export const Nav: FC = () => {
                 <i className="bi bi-check2-circle me-1"></i>Validation
               </a>
             </li>
+            {/*<li className="text-body mx-2 my-0 nav-item h5 pt-1">*/}
+            {/*  <a className="nav-link" href="/about">*/}
+            {/*    <i className="bi bi-info-circle me-1"></i>About*/}
+            {/*  </a>*/}
+            {/*</li>*/}
             <li className="text-body mx-2 my-0 nav-item h5 pt-1">
               {user ? (
                 <div className="mx-2 my-0 nav-item h5 pt-1">