+++ web/src/components/tables/sample-table.tsx
@@ -6,26 +6,42 @@ import { arraysToSampleList, sampleListToArrays } from '../../utils/sample-table
 import { addClassesToRows } from './hooks-callbacks';
 
 interface Props {
+  className?: string;
   data: Sample[];
   onChange?: (rows: Sample[]) => void;
   readOnly?: boolean;
   height?: number;
+  minRows?: number;
+  stretchH?: 'none' | 'all' | 'last';
 }
 /**
  * This table is meant to handle csv strings, so just pass in
  * the csv string and it will handle the rest
  */
-export const SampleTable: FC<Props> = ({ data, readOnly = false, onChange, height }) => {
+export const SampleTable: FC<Props> = ({ data, readOnly = false, onChange, height, minRows, stretchH, className }) => {
   // parse the list of objects into rows
   const rows = sampleListToArrays(data);
+  const ROW_HEIGHT = 23; // px
+
+  // compute table height based on number of rows
+  // or the minRows prop if it is provided
+  let tableHeight = rows.length * ROW_HEIGHT + 50;
+  if (minRows) {
+    tableHeight = minRows * ROW_HEIGHT + 50;
+  }
+
+  let tableClassName = 'rounded rounded-2';
+  if (className) {
+    tableClassName += ` ${className}`;
+  }
 
   return (
     <>
-      <div className="rounded rounded-2">
+      <div className={tableClassName}>
         <HotTable
           data={rows.length > 0 ? rows : [[]]}
-          stretchH="all"
-          height={height || 900}
+          stretchH={stretchH || 'all'}
+          height={height || tableHeight}
           readOnly={readOnly}
           colHeaders={true}
           dropdownMenu={true}
@@ -33,7 +49,7 @@ export const SampleTable: FC<Props> = ({ data, readOnly = false, onChange, heigh
             indicators: true,
           }}
           minCols={1}
-          minRows={50}
+          minRows={minRows || 50}
           contextMenu={[
             'row_above',
             'row_below',